Deflate the user interactions file
